+++ src/view/show_mode/editor/nodes/filternode.py
@@ -28,8 +28,8 @@ class FilterNode(Node):
         self._in_value_types: dict[str, DataType] = {}
         self._out_value_types: dict[str, DataType] = {}
 
-        # self.fsi = FilterSettingsItem(self, self.graphicsItem())
-        self.fsi = FilterSettingsItem(self._filter, self.graphicsItem())
+        self.fsi = FilterSettingsItem(self, self.graphicsItem())
+        # self.fsi = FilterSettingsItem(self._filter, self.graphicsItem())
         font: QFont = self.graphicsItem().nameItem.font()
         font.setPixelSize(12)
         self.graphicsItem().nameItem.setFont(font)
